+++ CHANGELOG.md
@@ -6,6 +6,31 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
 
+## [1.19.4] - 2025-03-16
+
+### Fix
+- Fixed panning bug on alternative map
+- Fixed missing airports on map
+
+
+
+## [1.19.3] - 2025-03-11
+
+### Changed
+- Updated FSE airport list to include the 20 new airports
+- Updated FSE aircraft list
+- Updated project dependencies
+
+
+
+## [1.19.2] - 2024-12-15
+
+### Changed
+- Updated FSE plane list (AW109SP and C408 - thx jonaseberle)
+- Updated project dependencies
+
+
+
 ## [1.19.1] - 2024-09-29
 
 ### Changed